(backend) loại bỏ code dư thừa sửa lại số lỗi, thực hiện business registration

+++ server/lms-server-web/src/main/java/com/hust/lms/streaming/service/impl/CategoryServiceImpl.java
@@ -40,6 +40,7 @@ public class CategoryServiceImpl implements CategoryService {
   @Override
   public void delete(UUID id) {
     Category category = this.categoryRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Category" , "id", id));
+    if (!category.getCourses().isEmpty()) throw new BadRequestException("Danh mục này không thể xóa");
     this.categoryRepository.delete(category);
   }
 }
